change divide method return double instead of int

+++ src/main/java/Calculator.java
@@ -16,7 +16,7 @@ public class Calculator {
         return num1 * num2;
     }
 
-    public int divide(int num1, int num2) {
+    public double divide(double num1, double num2){
         return num1/num2;
     }
 }
